Connected editor
