+++ packages/admin-panel/src/editor/EditModal.jsx
@@ -1,6 +1,6 @@
 /**
- * Tupaia MediTrak
- * Copyright (c) 2018 Beyond Essential Systems Pty Ltd
+ * Tupaia
+ * Copyright (c) 2017 - 2024 Beyond Essential Systems Pty Ltd
  */
 
 import React, { useEffect, useState } from 'react';
@@ -8,12 +8,12 @@ import keyBy from 'lodash.keyby';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Button, Dialog, DialogFooter, DialogHeader } from '@tupaia/ui-components';
-import { closeEditModal, editField, saveEdits } from './actions';
-import { getEditorState, getIsUnchanged } from './selectors';
-import { Editor } from './Editor';
+import { closeEditModal } from './actions';
+import { FieldsEditor } from './FieldsEditor';
 import { ModalContentProvider } from '../widgets';
 import { UsedBy } from '../usedBy/UsedBy';
 import { getExplodedFields } from '../utilities';
+import { withConnectedEditor } from './withConnectedEditor';
 
 const getFieldSourceToEdit = field => {
   const { source, editConfig = {} } = field;
@@ -30,76 +30,83 @@ const getFieldSourceToEdit = field => {
   return source;
 };
 
-export const EditModalComponent = ({
-  errorMessage,
-  isOpen,
-  isLoading,
-  onDismiss,
-  onEditField,
-  onSave,
-  recordData,
-  title,
-  fields,
-  FieldsComponent,
-  isUnchanged,
-  displayUsedBy,
-  usedByConfig,
-  dismissButtonText,
-  cancelButtonText,
-  saveButtonText,
-  extraDialogProps,
-}) => {
-  // key the fields by their source so we can easily find the field to edit. Use the exploded fields so that any subfields are placed into the top level of the array
-  const fieldsBySource = keyBy(getExplodedFields(fields), 'source');
+export const EditModalComponent = withConnectedEditor(
+  ({
+    errorMessage,
+    isOpen,
+    isLoading,
+    onDismiss,
+    onEditField,
+    onSave,
+    recordData,
+    title,
+    fields,
+    FieldsComponent,
+    isUnchanged,
+    displayUsedBy,
+    usedByConfig,
+    dismissButtonText,
+    cancelButtonText,
+    saveButtonText,
+    extraDialogProps,
+  }) => {
+    // key the fields by their source so we can easily find the field to edit. Use the exploded fields so that any subfields are placed into the top level of the array
+    const fieldsBySource = keyBy(getExplodedFields(fields), 'source');
 
-  // Files cannot be stored in redux (https://redux.js.org/style-guide/#do-not-put-non-serializable-values-in-state-or-actions)
-  // So instead we keep a map of fieldKey -> File, and store only the filename in redux.
-  const [files, setFiles] = useState({});
-  const handleSetFormFile = (inputKey, { fileName, file }) => {
-    setFiles({
-      ...files,
-      [inputKey]: file,
-    });
-    onEditField(inputKey, fileName);
-  };
-  useEffect(() => {
-    // Rely on the fact that opening/closing the modal clears `fields` to make sure that `files` gets wiped between forms.
-    setFiles({});
-  }, [fields]);
+    // Files cannot be stored in redux (https://redux.js.org/style-guide/#do-not-put-non-serializable-values-in-state-or-actions)
+    // So instead we keep a map of fieldKey -> File, and store only the filename in redux.
+    const [files, setFiles] = useState({});
+    const handleSetFormFile = (inputKey, { fileName, file }) => {
+      setFiles({
+        ...files,
+        [inputKey]: file,
+      });
+      onEditField(inputKey, fileName);
+    };
+    useEffect(() => {
+      // Rely on the fact that opening/closing the modal clears `fields` to make sure that `files` gets wiped between forms.
+      setFiles({});
+    }, [fields]);
 
-  const FieldsComponentResolved = FieldsComponent ?? Editor;
+    const FieldsComponentResolved = FieldsComponent ?? FieldsEditor;
 
-  return (
-    <Dialog onClose={onDismiss} open={isOpen} disableBackdropClick {...extraDialogProps}>
-      <DialogHeader onClose={onDismiss} title={title} />
-      <ModalContentProvider errorMessage={errorMessage} isLoading={isLoading}>
-        <FieldsComponentResolved
-          fields={fields}
-          isLoading={isLoading}
-          recordData={recordData}
-          onEditField={(fieldSource, newValue) => {
-            const fieldSourceToEdit = getFieldSourceToEdit(fieldsBySource[fieldSource]);
-            return onEditField(fieldSourceToEdit, newValue);
-          }}
-          onSetFormFile={handleSetFormFile}
-        />
-        {displayUsedBy && <UsedBy {...usedByConfig} />}
-      </ModalContentProvider>
-      <DialogFooter>
-        <Button id="form-button-cancel" variant="outlined" onClick={onDismiss} disabled={isLoading}>
-          {errorMessage ? dismissButtonText : cancelButtonText}
-        </Button>
-        <Button
-          id="form-button-save"
-          onClick={() => onSave(files)}
-          disabled={!!errorMessage || isLoading || isUnchanged}
-        >
-          {saveButtonText}
-        </Button>
-      </DialogFooter>
-    </Dialog>
-  );
-};
+    return (
+      <Dialog onClose={onDismiss} open={isOpen} disableBackdropClick {...extraDialogProps}>
+        <DialogHeader onClose={onDismiss} title={title} />
+        <ModalContentProvider errorMessage={errorMessage} isLoading={isLoading}>
+          <FieldsComponentResolved
+            fields={fields}
+            isLoading={isLoading}
+            recordData={recordData}
+            onEditField={(fieldSource, newValue) => {
+              const fieldSourceToEdit = getFieldSourceToEdit(fieldsBySource[fieldSource]);
+              return onEditField(fieldSourceToEdit, newValue);
+            }}
+            onSetFormFile={handleSetFormFile}
+          />
+          {displayUsedBy && <UsedBy {...usedByConfig} />}
+        </ModalContentProvider>
+        <DialogFooter>
+          <Button
+            id="form-button-cancel"
+            variant="outlined"
+            onClick={onDismiss}
+            disabled={isLoading}
+          >
+            {errorMessage ? dismissButtonText : cancelButtonText}
+          </Button>
+          <Button
+            id="form-button-save"
+            onClick={() => onSave(files)}
+            disabled={!!errorMessage || isLoading || isUnchanged}
+          >
+            {saveButtonText}
+          </Button>
+        </DialogFooter>
+      </Dialog>
+    );
+  },
+);
 
 EditModalComponent.propTypes = {
   errorMessage: PropTypes.string,
@@ -135,73 +142,8 @@ EditModalComponent.defaultProps = {
   extraDialogProps: null,
 };
 
-const mapStateToProps = state => ({
-  ...getEditorState(state),
-  isUnchanged: getIsUnchanged(state),
-  usedByConfig: {
-    usedBy: state.usedBy.byRecordId[state.editor.recordId] ?? [],
-    usedByIsLoading: state.usedBy.isLoading,
-    usedByErrorMessage: state.usedBy.errorMessage,
-  },
-});
-
 const mapDispatchToProps = dispatch => ({
   onDismiss: () => dispatch(closeEditModal()),
-  onEditField: (fieldKey, newValue) => dispatch(editField(fieldKey, newValue)),
-  dispatch,
 });
 
-const processRecordData = (recordData, fields) => {
-  if (Array.isArray(recordData) && Array.isArray(fields)) {
-    const [firstRecord] = recordData;
-    return fields.reduce((obj, field) => {
-      const value = field.bulkAccessor ? field.bulkAccessor(recordData) : firstRecord[field.source];
-      return { [field.source]: value, ...obj };
-    }, {});
-  }
-
-  return recordData;
-};
-
-const mergeProps = (
-  {
-    endpoint,
-    editedFields,
-    recordData,
-    usedByConfig: usedByConfigInMapStateProps,
-    initialValues,
-    ...stateProps
-  },
-  { dispatch, ...dispatchProps },
-  { onProcessDataForSave, usedByConfig: usedByConfigInOwnProps, ...ownProps },
-) => {
-  const usedByConfig = { ...usedByConfigInOwnProps, ...usedByConfigInMapStateProps };
-
-  return {
-    ...ownProps,
-    ...stateProps,
-    ...dispatchProps,
-    editedFields,
-    recordData: {
-      ...processRecordData(recordData, stateProps.fields),
-      ...initialValues,
-      ...editedFields,
-    }, // Include edits in visible record data
-    onSave: files => {
-      // If there is no record data, this is a new record
-      const isNew = Object.keys(recordData).length === 0;
-      let fieldValuesToSave = isNew ? { ...initialValues, ...editedFields } : { ...editedFields };
-      if (onProcessDataForSave) {
-        fieldValuesToSave = onProcessDataForSave(fieldValuesToSave, recordData);
-      }
-      dispatch(saveEdits(endpoint, fieldValuesToSave, isNew, files));
-    },
-    usedByConfig,
-  };
-};
-
-export const EditModal = connect(
-  mapStateToProps,
-  mapDispatchToProps,
-  mergeProps,
-)(EditModalComponent);
+export const EditModal = connect(null, mapDispatchToProps)(EditModalComponent);
