+++ packages/admin-panel/src/editor/index.js
@@ -10,3 +10,4 @@ export { BulkCreateButton } from './BulkCreateButton';
 export { EditModal } from './EditModal';
 export { reducer } from './reducer';
 export { CreateActionButton, ActionButton } from './ActionButton';
+export { withConnectedEditor } from './withConnectedEditor';
