+++ packages/admin-panel/src/surveys/SurveyEditFields.jsx
@@ -5,7 +5,7 @@
 
 import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
-import { Editor } from '../editor/Editor';
+import { FieldsEditor } from '../editor/FieldsEditor';
 import { useSuggestSurveyCode } from './useSuggestSurveyCode';
 import { useApiContext } from '../utilities/ApiProvider';
 import { useDebounce } from '../utilities';
@@ -33,7 +33,7 @@ export const SurveyEditFields = ({ fields, recordData, onEditField, ...restOfPro
   }, [suggestedCode]);
 
   return (
-    <Editor
+    <FieldsEditor
       fields={fields}
       recordData={recordData}
       onEditField={handleEditField}
